requirements.txt
atomicwrites 
attrs 
beautifulsoup4 
bs4==0.0.1
certifi 
chardet 
click 
colorama 
Flask 
idna 
iniconfig 
itsdangerous
Jinja2 
MarkupSafe 
packaging 
pluggy 
py 
PyJWT 
pymongo 
pyparsing 
pytest 
pytest-flask 
python-dotenv 
requests 
soupsieve 
toml 
urllib3 
Werkzeug 